DesarrolloWeb/requirements.txt
Django>=3.0,<4.0
django-filter == 22.1
django-widget-tweaks == 1.4.12
psycopg2>=2.8